+++ .iex.exs
@@ -4,7 +4,18 @@
 {:ok, tello_server_port} = Tello.CyberTello.port()
 |> IO.inspect()
 
-{:ok, supervisor, client, status_listener} = Tello.start(client: [ip: {127, 0, 0, 1}, port: tello_server_port])
+defmodule Receiver do
+  use Tello.Client.Receiver
+
+  def receive_message(data) do
+    IO.inspect("receives message from Tello: #{data}")
+  end
+end
+
+{:ok, supervisor, client, status_listener} =
+   Tello.start(
+      client: [ip: {127, 0, 0, 1}, port: tello_server_port, receiver: Receiver]
+    )
 |> IO.inspect()
 
 Tello.Command.enable(client)