+++ lib/tello/client/client.ex
@@ -5,7 +5,7 @@ defmodule Tello.Client do
 
   defmodule Tello.Client.State do
     @moduledoc false
-    defstruct [:socket, :tello_server]
+    defstruct [:socket, :tello_server, :receiver_module]
   end
 
   alias Tello.Client.State
@@ -15,17 +15,22 @@ defmodule Tello.Client do
 
   # Server (callbacks)
 
-  def start_link(uid: uid, arg: [ip: ip, port: port]) do
-    GenServer.start_link(__MODULE__, [ip: ip, port: port], name: :"#{__MODULE__}.#{uid}")
+  def start_link(uid: uid, arg: arg) do
+    GenServer.start_link(__MODULE__, arg, name: :"#{__MODULE__}.#{uid}")
   end
 
   @impl true
   def init(ip: ip, port: port) do
+    init(ip: ip, port: port, receiver: nil)
+  end
+
+  def init(ip: ip, port: port, receiver: receiver_module) do
     {:ok, socket} = :gen_udp.open(0, [:binary, active: true])
 
     state = %State{
       socket: socket,
-      tello_server: {ip, port}
+      tello_server: {ip, port},
+      receiver_module: receiver_module
     }
 
     {:ok, state}
@@ -48,8 +53,21 @@ defmodule Tello.Client do
   end
 
   @impl true
-  def handle_info({:udp, socket, ip, port, data}, state) do
-    Logger.info("Receives data from #{inspect(socket)}, #{inspect(ip)}:#{port}, data: #{data}")
+  def handle_info(
+        {:udp, socket, ip, port, data},
+        %State{receiver_module: nil} = state
+      ) do
+    Logger.debug("Receives data from #{inspect(socket)}, #{inspect(ip)}:#{port}, data: #{data}")
+
+    {:noreply, state}
+  end
+
+  def handle_info(
+        {:udp, _socket, _ip, _port, data},
+        %State{receiver_module: receiver_module} = state
+      ) do
+    receiver_module.receive_message(data)
+
     {:noreply, state}
   end
 end