+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## main
 
+
 - Added `Tello.Client.StatusListener` to receive status update from Tello.
+- Allow to attach custom receiver to `Tello.Client` to receive message form Tello, see `Tello.Client.Receiver` for more detail.
 
 ## V0.1.0